非同期のコメント・いいね機能の1つのシステムテストを2つのファイルに分割
